% Generated by roxygen2: do not edit by hand
\docType{package}
\name{miLAG-package}
\alias{miLAG}
\title{miLAG: Calculate and visualize microbial growth curve lag phase duration.}
\description{
\if{html}{\figure{milag-logo.png}{options: style='float: right' alt='logo' width='120'}}

A library that allows to estimate microbial lag phase duration based on user-provided data, the dataset must contain two columns. The accepted file formats are .csv and .txt, "calc_lag" is the main function of this library.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/bognabognabogna/microbial_lag_calculator/blob/main/shiny_app/lag_calulator/LAG_manual.pdf}
  \item \url{https://microbialgrowth.shinyapps.io/lag_calulator/}
  \item Report bugs at \url{https://github.com/bognabognabogna/microbial_lag_calculator/issues}
}

}
\author{
\strong{Maintainer}: Bogna J. Smug \email{bogna.smug@uj.edu.pl} (\href{https://orcid.org/0000-0001-9364-163X}{ORCID})

Authors:
\itemize{
  \item Monika Opalek \email{monika.opalek@doctoral.uj.edu.pl} (\href{https://orcid.org/0000-0003-1414-901X}{ORCID})
  \item Maks Necki
  \item Dominika Wloch-Salamon (\href{https://orcid.org/0000-0002-0040-1838}{ORCID})
}
}
\keyword{internal}
